requirements.txt
asgiref==3.5.0
autopep8==1.6.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
defusedxml==0.7.1
Deprecated==1.2.13
dj-rest-auth==2.2.4
Django==4.0.2
django-allauth==0.51.0
django-cors-headers==3.11.0
django-oauth-toolkit==2.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
idna==3.3
jwcrypto==1.3.1
oauthlib==3.2.0
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
python-decouple==3.6
python3-openid==3.2.0
pytz==2021.3
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.10
wrapt==1.14.1
